copr: Add vectorized Trim3Args.
Address comment. Remove Trim2Args because of duplicated for #7094.
